Disallow transition.{style,attr} getters.
There’s no good way to distinguish between whether the corresponding tween’s
end value is null or whether there is no corresponding tween. I mean, you could
return null for the former and undefined for the latter, but that seems unsafe.
Better to require the use of transition.{style,attr}Tween or transition.tween to
inspect the transition, and transition.{style,attr} be convenience functions.