Copy-on-write!
If the next node in a transition shares state with the previous node, as is
often the case due to transition method applying to the entire selection, then
we can reuse the state rather than creating separate state for each node.

This commit also starts to enforce restrictions on when transition state can be
modified; transition state should only be modified before the transition starts,
or in the case of delay or event listeners, before the transition is scheduled!

Related #24 #25 #27 #28.