Allow some listener changes after init.
You’re not allowed to modify start event listeners after the transition is
scheduled, so as to avoid changes during a start event (due to changes being
copy-on-write). However, you should still be allowed to change the OTHER event
listeners (interrupt and end) before the transition starts; in particular,
transition.remove should be allowed during a start event!