A specialized interpolator.
The target value for transition.{attr,style} should always be a number, a color
or a string; therefore, we can use a specialized interpolator that is more
efficient than d3.interpolator.

This commit also removes coercion of the target attribute or style value to a
string; that’s not done in selection.{attr,style}, so it should be necessary
here, and it’s more efficient if we can optimize numeric target values, too.