Protect against errors.
If a transition throws an error on start, we no longer repeatedly restart the
transition at every frame. Instead, we silently terminate the transition on the
next frame. If a transition throws an error on end, we no longer run the
transition forever at t=1. Instead, we silently terminate the transition at the
end of the current frame (or a subsequent frame, if an error is thrown by
another timer).

Fixes #48 #55.